Formulario de registro y modificaciones para conexión back-BD

+++ backend/src/main/java/com/TFG/app/backend/address/repository/AddressRepository.java
@@ -3,7 +3,6 @@ package com.TFG.app.backend.address.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.TFG.app.backend.address.entity.Address;
 
-public interface AddressRepository extends JpaRepository<Address, Long> {
-    // Additional custom queries can be defined here if needed
-    
+public interface AddressRepository extends JpaRepository<Address, Integer> {
+
 }
