+++ backend/src/main/java/com/TFG/app/backend/bank_account/repository/Bank_AccountRepository.java
@@ -3,6 +3,6 @@ package com.TFG.app.backend.bank_account.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.TFG.app.backend.bank_account.entity.Bank_Account;
 
-public interface Bank_AccountRepository extends JpaRepository<Bank_Account, Long> {
+public interface Bank_AccountRepository extends JpaRepository<Bank_Account, Integer> {
     
 }
