+++ backend/src/main/java/com/TFG/app/backend/bill/repository/BillRepository.java
@@ -3,8 +3,5 @@ package com.TFG.app.backend.bill.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.TFG.app.backend.bill.entity.Bill;
 
-public interface BillRepository extends JpaRepository<Bill, Long> {
-    // Additional custom queries can be defined here if needed
-
-    
+public interface BillRepository extends JpaRepository<Bill, Integer> {
 }
