+++ backend/src/main/java/com/TFG/app/backend/establishment/repository/EstablishmentRepository.java
@@ -3,7 +3,7 @@ package com.TFG.app.backend.establishment.repository;
 import com.TFG.app.backend.establishment.entity.Establishment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstablishmentRepository extends JpaRepository<Establishment, Long> {
+public interface EstablishmentRepository extends JpaRepository<Establishment, Integer> {
 
 
 }
