+++ backend/src/main/java/com/TFG/app/backend/periodic_spending/repository/Periodic_SpendingRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.TFG.app.backend.periodic_spending.entity.Periodic_Spending;
 
-public interface Periodic_SpendingRepository extends JpaRepository<Periodic_Spending, Long> {
+public interface Periodic_SpendingRepository extends JpaRepository<Periodic_Spending, Integer> {
   
 }
