+++ backend/src/main/java/com/TFG/app/backend/savings_goal/repository/Savings_GoalRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.TFG.app.backend.savings_goal.entity.Savings_Goal;
 
-public interface Savings_GoalRepository extends JpaRepository<Savings_Goal, Long> {
+public interface Savings_GoalRepository extends JpaRepository<Savings_Goal, Integer> {
     
 }
