+++ backend/src/main/java/com/TFG/app/backend/spending/repository/SpendingRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.TFG.app.backend.spending.entity.Spending;
 
-public interface SpendingRepository extends JpaRepository<Spending, Long> {
+public interface SpendingRepository extends JpaRepository<Spending, Integer> {
     
 }
