+++ backend/src/main/java/com/TFG/app/backend/street/repository/StreetRepository.java
@@ -3,7 +3,7 @@ package com.TFG.app.backend.street.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.TFG.app.backend.street.entity.Street;
 
-public interface StreetRepository extends JpaRepository<Street, Long> {
+public interface StreetRepository extends JpaRepository<Street, Integer> {
     
     
 }
