+++ backend/src/main/java/com/TFG/app/backend/type_periodic/repository/Type_PeriodicRepository.java
@@ -3,4 +3,4 @@ package com.TFG.app.backend.type_periodic.repository;
 import com.TFG.app.backend.type_periodic.entity.Type_Periodic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Type_PeriodicRepository extends JpaRepository<Type_Periodic, Long> {}
+public interface Type_PeriodicRepository extends JpaRepository<Type_Periodic, Integer> {}
